[linux] Removed path in the scheduler unused in Legacy library
The path in the scheduler covered by SYNCHRONIZATION_BY_NON_ZERO_THREAD
was never used and not compiled in the Legacy library since scheduler
is built without MFX_VA define.

The path was built for the scheduler included in HW plugins, but
(discussed with Oleg Nabiullin) should not be really used - scheduler
instance from the Legacy library should always be in use.

Change-Id: I353048f7ac1ab329e4b9485da6c10f0eb3636d42
Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>