improve cmake search for gtest libraries and headers
On some systems, gtest might be installed in the standard
system paths (e.g. /usr/lib, /usr/lib64, /usr/include).
In order to detect the system installed gtest include path,
we need to search for gtest/gtest.h instead of gtest.h and
adjust the custom /usr/local/include path hint accordingly.
All source files in this project include "gtest/gtest.h",
so this is ok.

For a local gtest compile, gtest does not support 'install'
so the user is forced to install the libraries/headers manually.
This is inconvenient.  So allow the user to specify the
paths to the local gtest compilation via environment vars.

Append and report the actual paths where the gtest libraries
and headers where found instead of hard-coding it to
/usr/local/...

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>