fix implicit narrowing conversions
Implicit narrowing conversions are not allowed in GCC c++0x/c++11
initializer lists.  And they are treated as errors in GCC 5.x.

Various list initializations of the tc_struct are attempting to stuff
floating-point and/or negative values into an unsigned type.  Thus,
change the type so that narrowing conversions are avoided in the list
initializers.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>