fix gnu inline semantics on EDIError
The inline semantics changed between GNU89 and C99. GCC 5
uses C99 by default.  Previous versions used GNU89.

In GNU89, an externally visible function is generated for 'inline'
functions.  But in C99, 'inline' is the opposite (i.e. no external
visiblity).

For maximum compatibility and to keep existing behavior, just force
the GNU89 inline semantics on EDIError regardless of the compilers
default standard.

This fixes an "inline function declared but never defined"
warning generated by GCC 5, which is ultimately treated as an
error due to the use of -Werror in our compiler flags.

See https://gcc.gnu.org/gcc-5/porting_to.html for more
details on the inline semantics change.

The other option would be to declare and define EDIError at the
same time (i.e. within the same TU).  However, it is not clear
to me what the original author was trying to achieve in splitting
the declaration from the definition of an inline function between
two separate TU's (to hide impl details? does it need external
visibility?).

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>