explicit MemID cast for comparison
MemID is unsigned, yet it appears most of the implementation here
assigns -1 to MemID variables to indicate an invalid id.  So in
order to do a comparison of -1 and MemID, we need to explicitly
cast -1 to a MemID.  Otherwise the compiler will complain about
comparisons between signed and unsigned integers.

This fixes compilation error on GCC 5.x:
  error: comparison between signed and unsigned integer expressions
  [-Werror=sign-compare]

Change-Id: I7d1fd651086942a2ac2d1bed9e8156f43062b72c
Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>