ts/msdk_ts: add WARNING_FLAGS to msdk_ts and msdk_gmock
Add WARNING_FLAGS to msdk_gmock and msdk_ts cmake targets.
This allows '-Wall -Werror' to be propagated into these
compile targets for strict compilation.

For now, don't add the warning flags to the msdk_ts_dll
target since msdk_ts_wrap.cxx still generates warnings.
It is not clear how this file is maintained.  That
is, there is a comment in that file stating that it
has been autogenerated.

Change-Id: Ib32f48459dbf71f9b6fb7bb986e8480d0ebcdc9a
Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>