dispatch: add cmake install prefix support
On Linux, set the CMAKE_INSTALL_PREFIX default to
/opt/intel/mediasdk. This overrides the cmake default and
also allows it to be manually overridden to something else.

Define MFX_MODULES_DIR and MFX_PLUGINS_DIR variables to
be relative to the install prefix (if not already defined
at a higher level).  Pass these variables to the compiler
command line so that the code can refer to them instead
of hardcoded paths.

Change-Id: I3c5ec60cd329e69511c001ded2da822189dac45a
Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>