ts/thread_safety: fix enum-compare compiler warning
Cast the 2nd operand to mfxU32 in the conditional operator
since the enum lives in the anonymous enum space.  The 3rd
operand gets promoted to mfxU32 by virtue of the cast in the
2nd operand.  The compiler can't resolve these anonymous
enums to the same type. So we need to help it.  [-Wenum-compare]

This bogus warning is possibly related to compiler bug:
  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=53524

Change-Id: Ie704cccf13e4631a7477b1f93da874695c670f41
Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>