sample_fei: fix narrowing conversion warning
For MaxFeiEncMFPNum, the #define deduces it to
an int of value 4.  Instead, declare MaxFeiEncMVPNum as a
const unsigned since it should never be a negative value,
it's constant, and is used in the context of unsigned
operations. [-Wnarrowing]

In C++, #defines should rarely be used this way.  Instead,
prefer const declarations.

Change-Id: I16bc3e8959fa4ff262ddfa1401bf659afab29229
Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>
(cherry picked from commit a5eddb00167ea85641f4e40f8309398860f265a3)