multi-transcode: add parentheses in conditional (behavior change)
Add parentheses around the first || condition.  This changes the
behavior.

Previously, if the first operand evaluated to true, then the
rest of the condition would be short-circuited (unevaluated)
and we enter the conditional body.

However, this change assumes that we only want to enter the body
only if we are using a libva backend AND using either one of the
listed VPPComp methods.  It seems this was the original intention
based on the presence of the LIBVA_SUPPORT preprocessor condition.

Change-Id: Id350ce6aa48c8919d96a5089844e53c2ea6ded68
Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>