[vp9e plugin] Eliminated all incorrect usages of MFX_CHECK_STS.
Issue-Id: MDP-31588

MFX_CHECK_STS was applied to function calls in many places. It's incorrect
to use this macro in this was (it references it's operand 2 times, so
function was called twice and only 2nd result is returned to upper stack).

Tested by beh_vp9e_encode_frame_async and beh_vp9e_init.

Signed-off-by: Stanislav Efremov <stanislav.efremov@intel.com>