MSVC2015 fix C4297 function assumed not to throw an exception but does
this is a fix for C++11 compatible compiler
By default, the compiler generates implicit noexcept(true) specifiers for user-defined destructors and deallocator functions
https://msdn.microsoft.com/en-us/library/84e2zwhh.aspx
Exception Specifications (throw) is deprecated in C++11 so we can't use it.
https://msdn.microsoft.com/en-us/library/wfa0edys.aspx
So must use noexcept(true)
http://en.cppreference.com/w/cpp/language/noexcept_spec
http://en.cppreference.com/w/cpp/language/destructor
Non-throwing function are permitted to call potentially-throwing functions.
Whenever an exception is thrown and the search for a handler encounters the outermost block of a non-throwing function, the function std::terminate is called
So we MUST fix this warning

Change-Id: I5babc04916b6609ad3623bde5e10c9dbb04b988c
Signed-off-by: Sergey Solodkov <Sergey.Solodkov@intel.com>