IPU-weaving + MSDK ADI/BOB
Refactor the sample_mondello code cater for both IPU-weaving
and GPU-weaving scenarios in the event of streaming from an
Interlace signal (by submitting V4L2_FIELD_ALTERNATE to the IPU).
In the process after obtaining the IPU-weaving signal from the
IPU, we may want to do ADI/BOB deinterlacing filters.

Added exception to handle IPU frame drop for the case GPU_WEAVING.

IPU does not guarantee if there is a frame drop. An exception
handling mechanism is needed to be done on the apps level to
fix the odd/even misalignment issue when it undergo VPP GPU weave
operation (This is the case where we submit V4L2_FIELD_ALTERNATE
to IPU).

Other changes includes fixing sample_mondello compilation issue
with respect to the following changes:
1] (Cherry-picked from 81899191e8cb31015d6733de364fd221d6ac2181)
2] (Cherry-picked from 6159b5d5334f0c304a7f509e8a36467fde5dbb84)
3] (Cherry-picked from 9b1c5ece8ca9bfc730eb8b77c329abb394952d35)

Change-Id: If5d8730c9cab5c45b1a1c9775996d7af89b9e2cf
Signed-off-by: Vincent Cheah Beng Keat <vincent.beng.keat.cheah@intel.com>