Using typeof() instead of decltype for Linux build that doesn't support C++11
We need this to remove dependency on C++11.
We have investigated that this way is the best way to add supporting of anonymous MSDK structures without using C++11.
There is no change in beh-tests building process, because it supports C++11 for Linux build.

Task: https://jira01.devtools.intel.com/browse/MDP-32285
Test: N/A

Change-Id: I43ee4a024a14371520d261e580f07e80a3ce0b27