[api][gamefly] added  AdaptiveMaxFrameSize to mfxExtCodingOption3
MDP-123 Game streaming enabling
MDP-32929 [RCR 1405599614]Scene change quality improvement

it is supposed to be mapped to DDI bAutoMaxPBFrameSizeForSceneChange
bAutoMaxPBFrameSizeForSceneChange if enabled, BRC may decide a larger P or B frame size than what UserMaxPBFrameSize
dictates when the scene change is detected. It may benefit the video quality.
The default value is 0 under which the normal BRC is applied.

Change-Id: Iaa4448202f6e661a6e9b1692732ee0a0185b7add
Signed-off-by: Sergey Solodkov <sergey.solodkov@intel.com>