[traces, encoders] Move MSDK structures comparing mechanism to traces
The reason of moving is optimization of feature initialization process.
Feature is now initializing in MFXTrace_Init() - not directly in Query.
This way of feature initialization has next advantages:
* No recreation of global types collection in Query
(performance of Query will not be affected)
* Feature is now placed in mfx_traces project,
so it can be easily used in any solution which uses mfx_traces
Also, feature is cropped from Open-Source now
(this will be addressed in separate change).

Issue: MDP-32285
Test: N/A

Change-Id: I641afad728ec766b39bc7eed28ffbe72d62de036