[builder] always set MSDK_*_ROOT variables
From Cmake documentation:
"If PARENT_SCOPE is present, the variable will be set in the scope above
the current scope. Each new directory or function creates a new scope.
This command will set the value of a variable into the parent directory
or calling function (whichever is applicable to the case at hand)."

We include *.cmake in the root cmake, so there is no upper level
directories to work in.

Plus added usage of this variables in `mfx_include_dirs()`.

Test: N/A
Issue: N/A

Change-Id: I1ed573443e671befcde3289539d585468dd9a605