[SMT][sample_common]: Fix memory leak in drm rendering mode
vaDestroyImage call in AllocImpl will cause the image and image buffer
destroyed completely by the driver, causing vaReleaseBufferHandle
in ReleaseResponse function failed to release the buffer handle.
In this case, reference count in libdrm is not decreased and memory is
not being freed at the end of the process.

Since vaDestroyImage will be called in ReleaseResponse function after
vaReleaseBufferHandle, vaDestroyImage in AllocImp should only be called
if we failed to acquire buffer handle.

Issue: HSDES-1504440417

Test: sample_multi_transcode

Change-Id: I81196f3466933e1db08a5b2a3bc7f996e5d0f0d5