[AVCe HW] Add MFE support
Frames from different streams/sessions are combined together as a
single dispatch of the ENC kernel (MF-ENC).This helps to increase
number of independent wave fronts executed in parallel, creating more
parallelism for ENC operation and thus increasing VME utilization.
As a result the EU array can process several frames within the same
execution time as a single frame. The logic of submit is modified
in such a way that all storage-items are allocated in the pool,
and dynamically move to a queue(list) depending on a submition time;
different algorithms can be implemented; currenly FIFO principle
is kept;

Needs to be built againts change in API:
I2548d05121906298871caa2b2afa353a1cd30701

Needs to be build against the folllowing change in Core
(as a new mfe search guid is used to correctly traverse cores):
I94cec5d2e94afe022addc97888b8ee24abe63986

To turn-on MFE, mfe must be enabled in sample, and appropriate
macro has to be defined during build process.

Issue: N/A
Test: Passrates of all tests should remain the same

Change-Id: Ib516512297068d6e202054f92c3129d68e279703