[samples] fix for output of statistics; add delta-time in ms
Adds include calculating time-deltas in ms; in addition, a bug with
incorrect dump of output statistics in Encode function has been fixed.
Also, if -stat-log is passed, but -stat param is missed, the number
of frames to be collected for statistics is initialized by max
number of frames to transcode.

Issue: no dump of output statistics; no defaults for stat. window
Test: manual

Change-Id: I41a5f79ef3bf12812f8c4d669696296093a7f260