[components] restore msdk core respect to components threading policies
Some components which use 2 entry points try to use different
threading policies for them. As a result component's GetThreadingPolicy()
is not called and essentially msdk core selects policy rather than
component. This highly complecates the code and introduces error
prone situation.

The primary reason for the above situation is definition of the
MFX_THREADING_(DEDICATED)_WAIT policy. It is defined, but actually
not implemented (and not needed). This patch slaughters definition
of these policies and restores msdk core respect of GetThreadingPolicy().

Special note for plugins. Plugins (even msdk pseudo ones) use single
entry point. As a result they are not subject for this problem and
we are free to change the code.

For the Open Source mediasdk libmfxsw_encode.cpp is stored in the
mdp_msdk-tools repository. The following change reflects this patch
for the noted file: Ic1e1592be357dfdc4dee007d0f017ef702f24654.

Issue: N/A
Test: manual

Change-Id: I6f8bd514582057f538a37c56d326c63be31908d3
Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>