[mfe-adapter] fix: to convert wait time from vm_ticks--> ms
After an issue with possible overflow of wait-time was fixed,
convertion from ticks to ms was missed when wait on cond variable
with a timer is invoked. In addition, redundant loop-exit condition
is removed as its already done in while condition.

Issue: too long waits in mfe-waiting loop due-to use ticks, not ms
Test: Manualy

Change-Id: I41463593425737ddad969b8f85e81f6dec1dd1f2