[SMT] to fix an issue with null pointer of a parent session
as an extra check was added in order to dump statistics in SMT,
there is an access to so-called parent session to get number of
frames. But, the pointer to a parent session is not checked
for null. Currently its fixed.

Issue: segfault if no parent session is provided
Tests: automatic

Change-Id: I206f6c44354827caf000ff8e81001497c8e9f439