[Library] Prototype prop files system on two projects (preparing)
*.props files system is needed to avoid random changes of build configurations
properties.
For example, one solved to add a new include directory to
AdditionalIncludeDirectories property. He added it in the build configuration
he work, but forgot to add to other configs too. The term "random changes"
means just this situation.
In this commit some common (defs) *.props files have been added.
This is the preparation for the next works.

Issue: MDP-36367
Test: Manual build

Change-Id: I822077b947fa13890d17ce9c61caf83458ccdcc8