[HEVCe] Add more checks for DirtyRect parameters
This change is aligning HEVCe DirtyRect parameters checks
with AVCe DirtyRect parameters checks.
"Test" field is filled by test that is under development now.
This test is an adaptation of beh_avce_dirty_rect just with changing
AVCe to HEVCe, so it includes same test-cases with same parameters,
as it is already done for AVCe.
All test-cases are passed, if this change is applied.

Issue: MDP-36823
Test: beh_hevce_dirty_rect./*

Change-Id: Ie050882c2324c5fde286a0d6de9dec62cca718a0