[VP9e-plugin] Updated DDI 0.966->0.968
Issue: VSMGWL-4850

Tests: Launched basic beh. tests to check backward compatibility for
Gen10 and Gen11 with the current BKC driver.

New DDI 0.968 includes DirtyRext params in caps structure.
DDI 0.968 is backward compatible with 0.966 on the driver side (all GENs).
When driver is updated to 0.968 it will not work with MSDK 0.966,
so MSDK should be updated first.

Change-Id: I2c3558479120477ab8fb780cb0d743b40a852963