[jpegd] Need software decode for YUV411 on BXT
The problem was with cases of JPEG decoding
YUV411 color format to RGB32. The VPG team
confirmed that the driver does not support
that case, in particular, SFC is not capable
to convert YUV411 to RGB32. See HSD-1805542620.
Hence I activate software fallback for this case
for APL platform.
Now all tests are passing. Checked for seed 640 as
was in test report with issues.
Validation logs after the fix:
\\jfspercbits001.amr.corp.intel.com\
shared_outgoing\MSDK_Project\rkazants\MDP-33292

Issue: MDP-33292
Test: jpegd_multiscan

Change-Id: I96ed8043c39b39e6c6e7dab8f2e5f24c6eaf3ec1