[scheduler] full mutex protection of scheduler threads
This reverts commit a03c3a66ed3e098560cfa666f7c1dd0240426a43.

Full mutex protection of scheduler threads suites 2 purposes. The
one is increased code robustness: it is less subject for race
conditions. Another is that we are one more step forward to
ever-sleeping threads which wake up only if they are signalled.

Update single-thread mode. Use mutex in the same was as it was
protect GetTask and MarkTaskCompleted calls.

Issue: MDP-18011
Test: h264d_conf, decoders_multiprocess, *_multiprocess_multisession

Change-Id: I6f64742233276b6d1f2bd67820d3c89b212e6be2
Signed-off-by: Sergey Solodkov <Sergey.Solodkov@intel.com>