[HEVCe] Fix rect (ROI, DirtyRect) boundaries in HEVCe
Driver expects a rect with the 'close' right bottom edge but
MSDK uses the 'open' edge rect, thus the right bottom edge which
is decreased by 1 below converts 'open' -> 'close' notation
Also, it was decided that we need to exclude invalid rects from the following
processing (particularly, transferring to the driver) and return INVALID_PARAM
in case of Query/Init. In case of RT-mode we dosen't return anything.

Issue: MDP-37921
Test: hevce_roi-lowpower-plugin_hevcehw,
hevce_dirtyrect-lowpower-plugin_hevcehw

Change-Id: Idffb82f33311a192d114d2517ebb86447f27a295