[HEVCe] fix alignment for NumMbPerSlice in AddTileSlices()
Added additional check for alignment for NumMbPerSlice
in case of SliceStructure == 2. The absence of this check leads to
MFX_WRN_INCOMPATIBLE_VIDEOPARAM when parameter for NumMbPerSlice
was set correctly.
If value of nLcuPerSlice will be modified inside this check,
warning will be generated in CheckVideoParam(), and aligned value
will be assigned to NumMbPerSlice.

Issue: MDP-37781
Test: beh_hevce_co2_max_slice_size_in_ctu/*

Change-Id: I4a5aed2b7671b4474004dbaa72df03dd9085f568