[beh] fix wrong NumSlice value in beh_hevce_encode_frame_async
This change is fixing wrong value, that test was setting for
NumSlices field. In case of test type = NSLICE_LT_MAX,
value for NumSlices was calculated like (maxSlices - 1),
but it is incorrect for VDEnc (SliceStructure = 2),
because for this encoder mode we can specify for NumSlice
only those values, which are composed of any number of rows,
but all must have the same size, except last one, which can be
smaller or equal to previous slices.
For example,if maxSlices = 8 (resolution is 720x480),
we cannot specify NumSlice = 7 to check test case NSLICE_LT_MAX.
The most closed correct value, that is less than max value,
will be 8/2 = 4. For odd values (i.e. maxSlices = 5) correct value
for NumSlice will be the result of a similar division, but with
rounding(5/2 ~= 3). Thus, it is most convenient to use CEIL_DIV.

Issue: MDP-38665
Test: beh_hevce_encode_frame_async./*

Change-Id: Ic1ef9f60d737a825677ccbb81e5d62302b537f9c