[H264e hw] Regression caused by Adaptive LTR ExtBRC with SCD
General crash of AVC encoder caused by SCD module being called
when it is not even required for task configured. Added conditions
for calling amtScd.init() so it will be only initialized when
sw external BRC is called, picture structure is progressive and
rate control is set to CBR or VBR. Also SCD cmdevice can be created
or SCD can be attached to existing device, destruction of it at close
only if created.

Issue: MDP-38722

Test: h264e_func

Change-Id: I7603d585ce7afa0a5b178fe32a2893d9ec320d97