[AVCe][SW] Fixed memory allocation
During SW fallback we allocate a big pool of
memory bsSize*numOfSliceEncs, where bsSize
depends on the frame size, numOfSliceEncs
equals to the number of processor's threads.
When 8K stream is being encoded on CFL platform
with 16 threads the amount of memory to allocate
is more than 2GB and ippMalloc function can not
allocate more than 2GB. To avoid this situation
allocate this memory pool by separate chunks
for each thread.

Issue: MDP-39535

Test: hevcd_to_h264e_long_play-plugin_hevcdhw

Change-Id: If38fa441a8f78f3be03588aff7976763e2b7c3a7