[beh] fix division by zero in hevce_co2_max_slice_size_in_ctu
This change is fixing division by zero exception.
It appeared due to division by cod2.NumMbPerSlice,
which value was set by rand(), so sometimes it was 0,
and that was a root-cause of exception.

Also, removed duplicated code and division
is replaced by CEIL_DIV macro (no impact,
just easier to read code).

Issue: MDP-39701
Test: beh_hevce_co2_max_slice_size_in_ctu./*

Change-Id: I6a8b1e6c914db7251e6f6550ca2f354bf80cbd35