[vpp][beh] Add gmock test for mirroring
Additionally, I have fixed a bug in Init function
that returned WRN_SKIPPED_FILTER return code
as if the feature is unsupported and ignored.
Also, there was a bug in mfx_hw_vpp
leading to the seg fault if CmCopyWrapper object
is not created in case of not loaded cmrt
run-time library.
All tests pass on SKL.

Test: beh_vpp_mirror
Issue: N/A

Change-Id: I3f7c42250b3a1c584fc422b8f5fef0bd6e71d95e
Signed-off-by: Roman Kazantsev <roman.kazantsev@intel.com>