[beh][hevce] Fix broken QpOffsets due to changes in p-pyramid feature
There was a problem on CNL VDEnc with beh_hevce_qp_offset:
for this platform, NumRefFrame = 2, NumRefActiveP = {2, 2, ... 2}.
It means that p-pyramid width must be = 2,
but test showed that values for QpOffsets indexes
are different from predicted for pyramid width = 2,
but equal to predicted for pyramid width = 3.
After investigation, we have figured out that
PPyerInterval is initializing before NumRefFrame has been
initialized. It means that it is always initializing to 3,
even if it must be initialized with another value due to
platform-specific limitation.
This change is fixing the place of PPyrInterval initialization
in code, moving it from MfxVideoParam::SyncVideoToCalculableParam()
to SetDefaults() function, right after NumRefFrame initialization.
Also, beh_hevce_qp_offset is updated to be aligned with new
p-pyramid feature behavior.

Issue: MDP-40313
Test: beh_hevce_qp_offset, beh_hevce_p_pyramid

Change-Id: I99654f23a5cca67439d4f29e92ced641425f148b