[mpeg2d] check if BITSTREAM_COMPLETE_FRAME is set
When BITSTREAM_COMPLETE_FRAME flag is set for input,
decoder trusts input too much, but later it works with
short input as it is end of stream.
In added check if input data size is less than picture
header and picture coding extension, it is dropped.

Issue: MDP-43023
Tests: issue content, custom validation to follow

Change-Id: I77c3abd100239dabcbb65ed1506eb66c61274905