genx: create a library holding genx kernels
As of now this library has 3 set of kernels:
 * copy_kernels
 * field_copy
 * h264_encode
The goal is to have all the kernels inside, i.e. add asc and mctf.

Side effect of this change is a drop if mediasdk library size (not stripped):
 * 44431992 before
 * 41268312 after

GitHub: e210fe4 ("genx: create a library holding genx kernels")
Issue: n/a
Tests: build

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>
Change-Id: I2c194c9201513e33540e021d0b32bcda4d9d5b2b