[VP9e/encoder] Fix for segmentation data init corruption
This fix overrides the previous one 9e097d62

This issue caused regression in encoding with segmentation after DDI
update. Local variable was sent to the driver and released in the end of
the scope before it started used. New DDI just changed memory alignment
on the stack and made this issue fire.

Issue: MDP-51012, VSMGWL-13530

Test: manual check

Change-Id: I0c3c53f730a3804086a394c06368b4bec51abe2e