[VPP] Fix for ASC CreateCmSurface2D crash for ADI
ASC crashes in ADI when ASC tries to create CmSurface for
downsampling operation. This issue appeared after the an
overload function CreateSurface2D definition was added in
cmrt_cross_platform.cpp,
https://git-ccr-1.devtools.intel.com/gerrit/#/c/76044/
this new overload receives mfxHDLPair as input but ASC
originally was casting the input from AbstractSurfaceHandle
to mfxHDLPair, and it internally was again casting it to
AbstractSurfaceHandle, but with the new overload the function
is interpreting it directly as mfxHDLPair which causes the
crash. By passing the input directly without mfxHDLPair
casting solves the issue, because it calls then the correct
CreateSurface2D function with AbstractSurfaceHandle input.

Issue: -
Test: Manual

Change-Id: I1588e48b2c00723bb9c50c710448153ebe6c2fd7