[H264e] ASC SWBRC bug fix caused by wrong surface input
SWBRC started to fail after commit to fix ADI_SCD crash
https://git-ccr-1.devtools.intel.com/gerrit/#/c/116150/
The modification on the commit did not take into account
that the handle for the input frame could be mfxHDLPair
or mfxHDL. mfxHDLPair is used in AVC encoder, while mfxHDL
is used by ADI and MCTF.
The fix consist of new overload functions for QueueFrame
and CreateCmSurface2D with surface input as mfxHDLPair,
in that way the surface input is interpreted correctly.
Also this fix removes the reinterpret_cast conversion in
SCD_Put_Frame for AVC HW encoder and let pass the
mfxHDLPair handle directly.

Issue: MDP-51319
Test: h264e_extbrc_adaptive_ltr, manual test ADI_SCD and MCTF.

Change-Id: I76f15ada20475dca063a25ccc3484cc6b3fa3a66