Set up bitDepth luma and chroma fields in the transcoder
According to MDP-33552, if VPP is used only for managing
memory then CopyPassThrough() method should be used instead
the driver. But this method isn't called because vpp.In is not
equal to vpp.Out. Values of BitDepthLuma and BitDepthChroma
fields of vpp.Out are not set.
Set up these fields in m_inParams.FrameInfo structure because
it's used in InitRenderParams() method when VPP is enabled.

Issue: MDP-36835
Test: vp9d_to_vp9e_opaque-plugin_vp9ehw_vp9dhw-lowpower

Change-Id: I8c2283f6b1270fe8b4b228c598039d8c8a15a532