[mfx_player] Set up an output bitDepth by FourCC
According to MDP-33552, if VPP is used only for managing
memory then CopyPassThrough() method should be used instead
the driver. But this method isn't called because vpp.In is not
equal to vpp.Out. Values of BitDepthLuma and BitDepthChroma
fields of vpp.Out are not set.
Added an initialization of the render bitDepth using the FourCC.

Issue: MDP-36835
Test: vp9d_to_vp9e_opaque-plugin_vp9ehw_vp9dhw-lowpower

Change-Id: I7631f6512e2e241a8f603234f1a1b7457624ac0a