[camera_pipe, aac_dec] Fix int-in-bool-context warnings breaking builds
Some code triggers int-in-bool-context GCC warnings that break Media SDK builds.
As a workaround developers suppress the warning in the builds completely, which is not in line with good code development practices.
To enforce good code development practices - even if the code is obsolete - need to fix it until it is removed from the code base.

Issue: MDP-61339, MDP-61340
Test: Compile with -Wall

Change-Id: Ifb1535c745f9771cfdbe6cea32dbd96c18a30cad