[ASF Parser] Fix a memset-elt-size GCC warning that breaks Linux builds
ASF Parser contains code that triggers memset-elt-size warning (enabled by -Wall), which is treated as error during builds.
Upon a quick look the code indeed is flawed: a smaller amount of memory than requested gets zeroed.
As a workaround developers suppress the warning in the entire builds completely, which is not in line with good code development practices.
To enforce good code development practices - even if the code is obsolete - need to fix the code until it is removed from the code base.

Issue: MDP-61352
Test: Compile with -Wall

Change-Id: Ic0a55cf5626b59cb95b091f1e75b94c605bba4c9