[mfx_player] Fix format-overflow warning breaking builds
There is code that triggers format-overflow GCC warning that break Media SDK builds.
As a workaround developers suppress the warning in the builds completely by specifying -Wno-format-overflow in FindGlobals.cmake (refer to the page
at Media SDK Wiki on building Media SDK on Linux). The code that triggers the warning has a potential buffer overflow vulnerability.
Such workarounds are not in line with good code development practices. To enforce good code development practices need to fix the code.

Issue: MDP-61362
Test: Compile with -Wall

Change-Id: I244d60d4b14a7ce975b791b9fbb75c8879b9b555