[AV1D] Change frameIdx in RefMap[] for non-displayed KEY_FRAME on Linux
Driver will return error in SetAvpPipeBufAddrParams because the active 
ref_list fails to be got in Av1ReferenceFrames::GetValidReference. The 
frameIdx is 0x7f for KEY_FRAME which surpass m_maxFrameIndex. So for Key
frame, also transfer the idx of RefMap[] to driver, and this idx is -1 
which is invalid frame and doesn't be used by driver side.

Issue: MDP-62700
Test: Allegro Rev1.1 clip Main_10bits_393_HighLevel_192x128_r5234.av1

Change-Id: Id7844190fca19646ddcb8a37d8d9d1f2d78e87a9