[AV1D] Fix access violation issue in bitstream parsing
MSDK will return error when try to get frame_id of the empty previous 
frame(This frame is deleted from dpb buffer if it won't be used as 
reference in further decoding,and this is normal situation). Create a 
variable to define the frame_id of every previous frame and use this to 
check current frame_id.

Issue: MDP-62700
Test: Allegro Rev1.1 clip Main_8bits_393_HighLevel_192x128_r5234.av1

Change-Id: I5e0ccec9311bb0e7150245821671ac5409384a2c