scheduler: enable linux scheduler on linux
On multi-core systems we can't effort to have too many active
threads which is the case with windows inherited scheduler. This
change lefts Windows scheduler for Windows and takes Linux
scheduler which we have long ago on github. In a way this change
brings status quo to embargo Linux.

Issue: VSMGWL-29920
Tests: build

Change-Id: I0e8c360fc36ba3f6a95bf3a60131e4f330177202
Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>