[HEVCe] Fix DPB handling issue when RollingI is enabled
Fix issue that pushing frames with temporal_layer_id > 0 to DPB when
Rolling I is enabled. Because Rolling I is only expected on the frames
which temporal_layer_id is 0, and its reference frame should be layer 0
as well.

Issue: -
Test: -
Github: 8f6bcd0 ([HEVCe] Fix DPB handling issue when RollingI is enabled)

Change-Id: Ibc56f5f0623ef1422dbef39e17f1681b694b05b2