Updating component test_system to revision da0b74acafb235f47e0693f2d03d62ab9d6560dd
[perf] Fix test type and WL ID in perf_density_1n_mp_smt_h264d_mixe test (#188)

It was decided to create a separate RPA test type for this test
instead of using the existing one (def-mp-mst).

Setting unique WL ID in such a way is applicable only to density_1n_mp_smt_h264d_mixe test,
so it must be done inside the related if-block.

All new WL ID and test type were created in RPA for this test.

Issue: MDP-65671
How tested: manually on Windows CFL:
\\nnlmdpfls01.inn.intel.com\mediasdk\atyugina