Updating component test_system to revision ca65359a4396a95ebc0c314abc02cb662f5ba0a3
[perf] Add sorting of SMT par-files parameters (#311)

The problem was that common::write_sample_multi_transcode_parfile()
function used to put parameters to SMT par-file in a random order.
This resulted in the same set of parameters but in a different order
from run to run, which made debugging more difficult.

The solution is to add sorting of SMT parameters to this function.

Issue: MDP-67002
How tested: manual run of several tests based on different scripts
that use write_sample_multi_transcode_parfile function. The results are
as expected. Logs collected on TGL Win: \\nnlmdpfls01\mediasdk\atyugina