Updating component test_system to revision fe6413c535397f6c54e7431ab772ffd6233add2e
[perf] Change RPA usage model for N->N perf test + minor fixes (#318)

perf_density_nn_mp_smt_hevc10d_hevc10e test can't be in the
"Multisession 1N with predefined output" RPA usage model
because it has only one unique output and must be placed in
the "Multisession" RPA usage model.

+ Add out_scan field to workloads_density_1n_mp_smt_hevc10d_hevc10e.yaml
as out_scan is a mandatory field for RPA though it is not necessary for
this test.

Issue: MDP-67002
How tested: manually locally